refactor: companions -> companion
