+++ src/main/java/com/tikitaka/triptroop/companion/domain/type/OpenStatus.java
@@ -1,0 +1,5 @@
+package com.tikitaka.triptroop.companion.domain.type;
+
+public enum OpenStatus {
+    OPEN, COMPLETED
+}
