--- src/main/java/com/tikitaka/triptroop/companions/domain/type/OpenStatus.java
@@ -1,5 +1,0 @@
-package com.tikitaka.triptroop.companions.domain.type;
-
-public enum OpenStatus {
-    OPEN, COMPLETED
-}
